// src/generated/api/admin/models/adminMealItemLabelingItemResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aller Meal API
 * Aller Meal Release 1 public, member, and admin API contract.
 * OpenAPI spec version: 1.0.0
 */
import type { AdminMealItemLabelingItemResponseMealType } from "./adminMealItemLabelingItemResponseMealType";
import type { AdminMealItemLabelingItemResponseStatus } from "./adminMealItemLabelingItemResponseStatus";

export interface AdminMealItemLabelingItemResponse {
  mealItemId?: string;
  mealId?: string;
  schoolId?: string;
  schoolName?: string;
  mealDate?: string;
  mealType?: AdminMealItemLabelingItemResponseMealType;
  name?: string;
  displayOrder?: number;
  status?: AdminMealItemLabelingItemResponseStatus;
  createdAt?: string;
  updatedAt?: string;
}