// src/generated/api/admin/models/adminFailedCollectionJobItemResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aller Meal API
 * Aller Meal Release 1 public, member, and admin API contract.
 * OpenAPI spec version: 1.0.0
 */
import type { AdminFailedCollectionJobItemResponseMealType } from "./adminFailedCollectionJobItemResponseMealType";

export interface AdminFailedCollectionJobItemResponse {
  collectionJobId?: string;
  schoolId?: string;
  mealDate?: string;
  mealType?: AdminFailedCollectionJobItemResponseMealType;
  responseTimeMillis?: number;
  collectionDurationMillis?: number;
  rawObjectId?: string;
  failureCode?: string;
  failureMessage?: string;
  createdAt?: string;
  updatedAt?: string;
}