// src/generated/api/admin/models/adminRecollectionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aller Meal API
 * Aller Meal Release 1 public, member, and admin API contract.
 * OpenAPI spec version: 1.0.0
 */
import type { AdminRecollectionResponseStatus } from "./adminRecollectionResponseStatus";

export interface AdminRecollectionResponse {
  originalCollectionJobId?: string;
  collectionJobId?: string;
  status?: AdminRecollectionResponseStatus;
  duplicate?: boolean;
}